+++ packages/rich-text-editor/src/RichTextEditor/RichTextEditor.tsx
@@ -112,7 +112,7 @@ export const RichTextEditor: React.VFC<RichTextEditorProps> = props => {
       <div className={styles.editorWrapper}>
         {controls && (
           <Toolbar aria-controls={editorId} aria-label="Text formatting">
-            {Object.values(controlMap).map((section, index) => (
+            {controlMap.map((section, index) => (
               <ToolbarSection key={index}>
                 {section.map((controlConfig, controlKeyIndex) => (
                   <ToggleIconButton